Update share-single Documentation
This PR aims to update the docs in order to accurately reflect the steps needed to take in order to actively use the Share.singleShare method.

Previously, the docs noted that the method to use was simply `share`. This was causing console warnings that were resolved when checking the source code for a matching method and noticing that one did not exist.