[android] ensure instagram stick is alway granted read permission
There were certain code paths previously where the instagram sticker was
not flagged with FLAG_GRANT_READ_URI_PERMISSION, which meant that
Instagram was unable to open the sticker. This moves the logic around
granting permission next to the logic about adding the sticker to the
intent. It is duplicated, but it was going to be duplicated _anyway_. At
least if it's here, we're only looking into the options deeply one time.

fixes #855